pyproject.toml
[project]
name = "gree-cli"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "greeclimate>=2.1.0",
    "pydantic>=2.11.7",
    "typer-slim>=0.16.0",
]

[project.scripts]
gree = "gree_cli.__main__:app"

[tool.uv]
package = true
